% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{calc.gt.bench.ss}
\alias{calc.gt.bench.ss}
\title{Compute genotype benchmarking summary statistics}
\usage{
calc.gt.bench.ss(plot.dat, ss.prefix, summarize.all = TRUE, trio.mode = FALSE)
}
\arguments{
\item{plot.dat}{List of benchmarking data processed by \code{\link[=get.gt.bench.plot.data]{get.gt.bench.plot.data()}}}

\item{ss.prefix}{Character prefix for summary statistic descriptions}

\item{summarize.all}{Should a final series of summary rows be added to reflect
the overall summary statistics of all data in \code{plot.dat}? [default: TRUE]}

\item{trio.mode}{Is the input data from parent-child trios? [default: FALSE]}

\item{seealso}{\code{\link[=get.gt.bench.plot.data]{get.gt.bench.plot.data()}}}
}
\value{
data.frame of summary statistics for various strata within \code{plot.dat}
}
\description{
Calculate summary metrics from output of \code{\link[=get.gt.bench.plot.data]{get.gt.bench.plot.data()}}
}
