% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{calc.gt.bench.stats}
\alias{calc.gt.bench.stats}
\title{Calculate genotype benchmarking statistics}
\usage{
calc.gt.bench.stats(
  bench.dat,
  vc = NULL,
  vsc = NULL,
  freq = NULL,
  zygosity = NULL,
  key.cols = 1:5
)
}
\arguments{
\item{bench.dat}{A list of data.frames of genotype benchmarking data loaded
with \code{\link[=load.gt.benchmark.tsvs]{load.gt.benchmark.tsvs()}}}

\item{vc}{Optional character vector of variant classes for subsetting
prior to calculations}

\item{vsc}{Optional character vector of variant subclasses for subsetting
prior to calculations}

\item{freq}{Optional character vector of frequency categories for subsetting
prior to calculations}

\item{zygosity}{Optional character vector of zygosities to consider for calculations}

\item{key.cols}{Numeric vector of column indexes to treat as
unique keys [default: 1:5]}
}
\value{
List of data.frames of benchmarking summary statistics amenable with
downstream plotting functions
}
\description{
Calculate the benchmarking statistics necessary for plotting sample-level
genotype benchmarking
}
\seealso{
\code{\link[=load.gt.benchmark.tsvs]{load.gt.benchmark.tsvs()}}
}
