% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{calc.gt.bench.vpg}
\alias{calc.gt.bench.vpg}
\title{Calculate variants per genome for genotype benchmarking}
\usage{
calc.gt.bench.vpg(plot.dat)
}
\arguments{
\item{plot.dat}{List of benchmarking data processed by \code{\link[=get.gt.bench.plot.data]{get.gt.bench.plot.data()}}}
}
\value{
integer of the average number of variants per genome
}
\description{
Compute total variants per genome from all elements in the output of \code{\link[=get.gt.bench.plot.data]{get.gt.bench.plot.data()}}
}
