% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics.R
\name{classify.gt.zygosity}
\alias{classify.gt.zygosity}
\title{Zygosity classification}
\usage{
classify.gt.zygosity(gt)
}
\arguments{
\item{gt}{Character vector of GT values formatted according to VCF spec}
}
\value{
Character vector of zygosity classifications for each value in
input \code{gt}, either \code{het} or \code{hom}
}
\description{
Parse a VCF-style GT character vector and classify the zygosity of each GT
}
\details{
Ignores null/missing alleles and reports \code{het} if there are two
or more unique alleles observed, otherwise reports \code{hom}
}
