% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{clean.breaks}
\alias{clean.breaks}
\title{Clean size or AF break labels}
\usage{
clean.breaks(vals)
}
\arguments{
\item{vals}{A character vector of break values to be parsed}
}
\value{
A named vector of numeric values corresponding to \code{vals}, where
element names correspond to the original entries in \code{vals}
}
\description{
Converts size or AF breaks encoded from characters to numeric values
}
