% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{create.cancer.colors}
\alias{create.cancer.colors}
\title{Generate cancer colors}
\usage{
create.cancer.colors(
  cancers,
  n.shades = 2,
  saturation.range = c(0.3, 0.95),
  value.range = c(0.35, 1),
  period = NULL,
  plot.colors = FALSE,
  plot.palettes = FALSE
)
}
\arguments{
\item{cancers}{Character vector of cancers, in order to be mapped onto rainbow}

\item{n.shades}{Number of shades (and highlights) appended to each cancer
type-specific palette [default: 2]}

\item{saturation.range}{Values passed to \code{\link[RLCtools:categorical.rainbow]{RLCtools::categorical.rainbow()}} [default: c(0.3, 0.95)]}

\item{value.range}{Values passed to \code{\link[RLCtools:categorical.rainbow]{RLCtools::categorical.rainbow()}} [default: c(0.35, 1)]}

\item{period}{Value passed to \code{\link[RLCtools:categorical.rainbow]{RLCtools::categorical.rainbow()}} [default: \code{length(cancers) / 3}]}

\item{plot.colors}{Should a cancer color diagnostic plot be generated? [default: FALSE]}

\item{plot.palettes}{Should a cancer palette diagnostic plot be generated? [default: FALSE]}
}
\value{
two-element list, containing \code{cancer.colors} and \code{cancer.palettes}
}
\description{
Automatically generate colors and palettes for all cancer types in G2C
}
\details{
A value for "control" should not be provided in \code{cancers};
this is automatically appended with preset color values
}
\seealso{
\code{\link[RLCtools:hsv.palette]{RLCtools::hsv.palette()}}, \code{\link[RLCtools:categorical.rainbow]{RLCtools::categorical.rainbow()}}
}
