% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{explode.by.cancer}
\alias{explode.by.cancer}
\title{Explode variable by cancer}
\usage{
explode.by.cancer(cancer, x = NULL)
}
\arguments{
\item{cancer}{Character vector with one element per participant, where each
element is a semicolon-delimited list of cancer diagnoses for that participant}

\item{x}{(Optional) vector of values to be exploded}
}
\value{
List of two vectors:
\itemize{
\item \code{cancer} : exploded version of input cancer labels, \code{cancer}
\item \code{x} : exploded version of input target vector, \code{x}
}
}
\description{
Expands values of a single variable such that there is a one-to-one mapping
between cancer diagnoses and the target variable of interest. In other words:
this function repeats values of a target variable in situ for individuals with
multiple cancer diagnoses
}
\details{
If no \code{x} is provided, only the cancer labels will be exploded
}
