% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.famfile}
\alias{load.famfile}
\title{Read & clean .fam file}
\usage{
load.famfile(in.fam, header = FALSE)
}
\arguments{
\item{in.fam}{Path to input .fam file}

\item{header}{Does \code{in.fam} have a header line? [default: FALSE]}
}
\value{
six-column \code{data.frame}
}
\description{
Reads an input .fam file and enforces PLINK standard six-column format
}
\seealso{
\url{https://www.cog-genomics.org/plink/1.9/formats#fam}
}
