% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.sample.qc.df}
\alias{load.sample.qc.df}
\title{Read sample QC manifest}
\usage{
load.sample.qc.df(tsv.in)
}
\arguments{
\item{tsv.in}{Path to QC .tsv}
}
\description{
Load & clean a sample intake QC manifest as a dataframe
}
