% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{plot.all.gt.bench.strata}
\alias{plot.all.gt.bench.strata}
\title{Plot all genotype benchmarking comparisons}
\usage{
\method{plot}{all.gt.bench.strata}(
  bench.dat,
  out.prefix,
  set.colors,
  ref.title = "external benchmarking",
  metric.name = NULL,
  title.preps = c("for", "in")
)
}
\arguments{
\item{bench.dat}{A list of data.frames of genotype benchmarking data loaded
with \code{\link[=load.gt.benchmark.tsvs]{load.gt.benchmark.tsvs()}}}

\item{out.prefix}{Path and filename prefix for all plots}

\item{set.colors}{Vector of colors for each sub-element in \code{plot.dat}
[default: greyscale]}

\item{ref.title}{Character name for reference/external benchmarking dataset
[default: "external benchmarking"]}

\item{metric.name}{Name of metric being evaluated}

\item{title.preps}{Prepositions to add in title [default: c("for", "in")]}
}
\description{
Main wrapper to plot a single comparison (PPV or sensitivity) across all strata
}
\seealso{
\code{\link[=load.gt.benchmark.tsvs]{load.gt.benchmark.tsvs()}}, \code{\link[=plot.gt.bench]{plot.gt.bench()}}
}
