% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{plot.gt.bench}
\alias{plot.gt.bench}
\title{Plot genotype benchmarking results}
\usage{
\method{plot}{gt.bench}(
  plot.dat,
  strata.names = NULL,
  set.colors = NULL,
  metric.name = NULL,
  title = NULL,
  af.cutoff = NULL,
  strata.space = 0.15,
  bar.hex = 0.6,
  whisker.hex = 0.035,
  parmar = c(1.8, 2.55, 1.8, 2.5)
)
}
\arguments{
\item{plot.dat}{List of genotype benchmarking data.frames loaded by
\code{\link[=get.gt.bench.plot.data]{get.gt.bench.plot.data()}}}

\item{strata.names}{Vector of names for each element in \code{plot.dat}
[default: use \code{names(plot.dat)}]}

\item{set.colors}{Vector of colors for each sub-element in \code{plot.dat}
[default: greyscale]}

\item{metric.name}{Name of metric being evaluated}

\item{title}{Title for plot}

\item{af.cutoff}{Allele frequency cutoff to distinguish between common and
rare variants; will be used to add second row to bottom X-axis with AF bins}

\item{strata.space}{Space between strata specified in user X units [default: 0.15]}

\item{bar.hex}{Relative horizontal expansion parameter for the median bars
and the outer "whiskers"; must be bounded on [0, 1] [default: 0.6]}

\item{whisker.hex}{Relative horizontal expansion for median whiskers [default: 0.035]}

\item{parmar}{Value of \code{mar} passed to \code{par()}}
}
\description{
Plot a user-specified set of sample-level genotype benchmarking strata
}
