% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{read.compressed.distrib}
\alias{read.compressed.distrib}
\title{Import a compressed summary distribution}
\usage{
read.compressed.distrib(tsv.in, key.cols = 1:2)
}
\arguments{
\item{tsv.in}{Path to input .tsv with compressed distribution}

\item{key.cols}{Column indexes to be treated as categorical key values [default: 1-2]}
}
\value{
A two-element list of \code{df}, a dataframe of the distribution encoded
in \code{tsv.in}, and \code{breaks}, a numeric vector corresponding to the column-axis
break values
}
\description{
Import a precomputed compressed summary distribution from a .tsv
}
\seealso{
\link{clean.breaks}
}
