% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{scatterplot.point.params}
\alias{scatterplot.point.params}
\title{Determine point parameters for a scatterplot}
\usage{
scatterplot.point.params(
  n,
  cex.floor = 1000,
  cex.ceiling = 100000,
  cex.start = 0.6,
  cex.end = 0.05,
  pch.cutoff = 10000,
  pch.big = 1,
  pch.small = 19,
  alpha.floor = 100000,
  alpha.ceiling = 1000000,
  alpha.start = 1,
  alpha.end = 0.01
)
}
\arguments{
\item{n}{Total number of points to be plotted}

\item{cex.floor}{Minimum sample size for dynamic \code{cex} scaling}

\item{cex.ceiling}{Maximum sample size for dynamic \code{cex} scaling}

\item{cex.start}{Largest value of \code{cex}; used when \code{n <= cex.floor}}

\item{cex.end}{Smallest value of \code{cex}; used when \code{n >= cex.ceiling}}

\item{pch.cutoff}{Maximum value of \code{n} before switching from \code{pch.big} to \code{pch.small}}

\item{pch.big}{Value of \code{pch} to use when \code{n <= pch.cutoff}}

\item{pch.small}{Value of \code{pch} to use when \code{n > pch.cutoff}}

\item{alpha.floor}{Minimum sample size for dynamic \code{alpha} scaling}

\item{alpha.ceiling}{Maximum sample size for dynamic \code{alpha} scaling}

\item{alpha.start}{Largest value of \code{alpha}; used when \code{n <= alpha.floor}}

\item{alpha.end}{Smallest value of \code{alpha}; used when \code{n >= alpha.ceiling}}
}
\value{
three-element list of \code{cex}, \code{pch}, \code{alpha}
}
\description{
Automate selection of \code{pch}, \code{alpha}, and \code{cex} for points in a scatterplot
depending on the total number of points to be plotted
}
