% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfqc.R
\name{load.gt.benchmark.tsvs}
\alias{load.gt.benchmark.tsvs}
\title{Load genotype benchmarking data}
\usage{
load.gt.benchmark.tsvs(tsvs.in, set.names, key.cols = 1:5, trio.mode = FALSE)
}
\arguments{
\item{tsvs.in}{Character vector of paths to .tsv inputs}

\item{set.names}{Character names for each .tsv in \code{tsvs.in}}

\item{key.cols}{Numeric vector of column indexes to treat as
unique keys [default: 1:5]}

\item{trio.mode}{Is the input data from parent-child trios? [default: FALSE]}
}
\value{
List of one data.frame per element in \code{tsvs.in}
}
\description{
Load & postprocess a set of genotype benchmarking .tsvs
}
